backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
sentence-transformers
langchain
chromadb
pdfplumber
watchdog
pygetwindow 
pywin32
darts
openai-whisper
psutil
langchain-community
together
langchain-community 
langchain-huggingface 
langchain-chroma 
langchain-together
moviepy 
pyPDF2 
python-docx